rename salary files
